Initial oanda broker implementation
Still need to implement the rest of the broker
interface and test. This is a good start

Had to change broker interface to add CloseTrade.
Right now we just create an order in the opposite
direction of our existing order.

When an order is executed Oanda creates a Trade.
Trades should be modified for closed rather than
creating a new order in the opposite direction.
I still need to add ModifyTrade and implement things
in the defaultbroker

I also removed all the auth token values from
the RPCs. The token will live inside each instance
of OandaBroker and that way we need not pass tokens
with each request